Fix bug on diff download
In diff mode, we can apply several diffs to an extract, and stop applying diffs
before reaching the latest version (if num_diffs > 30 or last two diffs have
the same timestamp). In this case, the previous code was incorrectly replacing
state.txt init_osmosis_diff(), instead of keeping it, to restart applying diffs
from this point.

Now, we skip init_osmosis_diff() if extract was updated.