oh_sanitizer: Fix handling of the comma to pass the tests
Ideally we would want a space after commas for cosmetic reasons when
they are a rule separator but in no other cases (so not in time spans,
weekday lists, etc.).
However getting the parser to correctly distinguish the two cases is
hard / impossible and commas are very rarely used as a rule separator.
So do the right thing for the most common case which fixes the tests.
Add a couple more tests to show and document the edge cases.

Signed-off-by: Francois Gouget <fgouget@free.fr>