oh_sanitizer: A time span may not have an end time
The opening_hours specification allow timespans to not have an end time
and this is widely used. It describes '08:00' as representing a
theoretical one minute long opening time.
This makes strings like '9 12' ambiguous: it could be "09:12',
'09:00-12:00' or '09:00,12:00'. Thus spaces are only recognized as a
timespan separator if the first timespan has an end time such as in
'09:00-1:00 18:00'.

Signed-off-by: Francois Gouget <fgouget@free.fr>