function overloading not possible in Python
is is not possible to have multiple functions sharing the same name and differing in the parameters.

Removed the not working one and provided an implementation using default parameters to provide the intended signature. Slicing with None as the end parameter works the same way as leaving it out.

Also replaces the test by checking for identity instead of equality to speed it up a bit and comply with PEP8.