Add warnings for parking:condition:<side>
`parking:condition:<side>` only makes sense for some values of
`parking:lane:<side>`. This commit adds two checks:

* For `parking:lane:<side>=separate`, warn that conditions muse be
mapped on the separately mapped object instead

* For `parking:lane:<side>` values that indicate no parking is
permitted, warn that conditions should not be set either

I have left out the `parking:lane:<side>` value `marked` here, because
its use is unfamiliar to me. For the rest of these checks the usage is
according to what is documented on the wiki.