Use `tt` instead of `t` in the subclass
Most of the time it'll be the same for the existing warning (with class 31600), only in rare cases where you have i.e. `|merge_to_right;through`, `t` would be merge_to_right, but `tt` is `merge_to_right;through`. Since the other warning is new and uses both `elif` and a `2` as prefix, using `tt` shouldn't result in non-unique values I think.