Avoid duplicate warning for 1-member multipolygon (#1352)
* Prevent duplicate warning

* Raise priority to match disabled test

The disabled test (other commit) has priority level "2". This test would trigger at level "3". Raised to make sure a warning is still shown at priority level 2

* Prevent trap

Prevent warning suggesting to make it a simple polygon if a wrong role is present, e.g. when you have a multipolygon with two members:
`outer`: a way
`inne`: a way inside the outer with a typo in the role

* Undo reshift, replace by err_roles check

As requested

* Typo