Add bad junction=roundabout check
Roundabouts per (OSM) definition must have the right of way. If not, they should be tagged as junction=circular.
Alternatively, it can also be a roundabout where the (i.e.) traffic signals are located on the connection between the roundabout and the incoming way. In this case, the (i.e.) traffic_signals should be moved to the incoming way only.
This check locates roundabouts that contain either a `highway=give_way` or `highway=stop` sign, or a `highway=traffic_signals`