Accept whitelisted values in ;-separated combinations
Accept whitelisted values also if they occur in a `;`-separated list. (Only for "open" keys; not for "closed" keys with fixed values like `oneway`)

Note, the easy way out would be:
if any(x in tags[k].split(";") for x in self.exceptions_open[k]):
but that would fail if a whitelisted and 'bad' entry are combined.
Given that this part of the code is only ran in rare cases, I guess it doesn't harm to have the check as detailed as possible.