Fix capture group counting
Everything in [...] is counted as one capture group.
Since booleanExpression is also used for what's inside the [...] blocks if there's an operator in the block, it was triggered too often.

Instead I moved the counter to exitAttribute_selector, which handles the exit of a [...] block (either booleanExpression or a predicate_simple)