Add warning for repeated conditions
This adds a warning when a condition is repeated, overruling (in case of ;-separated conditions) or at least making redundant (in case of AND-separated conditions) the preceding condition.

In addition it should fix getting a duplicate marker for such cases, in case there's a warning for the condition.

Whitelists *:conditional from TagFix_DuplicateValue as the new check is more thorough