Fix rollercoasters
The roller coaster exception in ImportDatabase_Prepare is wrong.

attraction=roller_coaster does refer to the area where the entire roller coaster is located (see wiki: https://wiki.openstreetmap.org/wiki/Tag:attraction%3Droller_coaster ). Consequently, it must be a polygon (or a node).

For the track, the tag roller_coaster=track (https://wiki.openstreetmap.org/wiki/Tag:roller_coaster%3Dtrack) has to be used. The wiki's both explicitly state that this tag should not be combined with attraction=roller_coaster. The track is of course not a polygon, even if it's a closed loop.