Add tidal_channel as a valid waterway
Per the wiki, the tidal_channel exists outside of the coastline. Hence, coastline_sinkhole will not contain the end point of a stream connected to such a channel.

Fixing it this way looked easier than adding it to coastline_sinkhole.

Also removes an unused (and probably wrongly defined) column.

See #2086