Fix #2179 - exclude multipolygon inners
For multipolygon relations, exclude the `inner` area. For other relations (mostly `site` or `building`) retain the old behaviour.

The conversion of ST_Contains to ST_Covers is needed because exact overlaps of ways with outers of multipolygons aren't detected otherwise. (Other solution would be to add a buffer like with other relations)