Exclude self-intersecting cycleways
Exclude cycleways that return to itself (in a P-shaped way). This fixes an incorrect warning about unconnected cycleways at the self-intersection point.

Technically enabled for all highways, but I guess the number of ways with level < 3 that self-intersect is negligible.

By using this instead of ST_IsSimple we make sure we still find cycleways that aren't connected on the "other end" too