Detect incorrect bridge/tunnel-highway connection
Implements a check to detect (non-bridge/tunnel) highways that are connected to the (non-start/end) nodes of tunnels or bridges.

For tunnels I only enabled it for 'higher priority' ways due to many false positives with e.g. footways etc.

For bridges I disabled boardwalks as hiking routes apparently sometimes require people to step off a boardwalk (which is typically not too high).

In all cases steps and anything that may be indoor/building related are disabled

Distinguishes between through-connections and end-node connections as requested by ComradeRamen