Add plugin for detecting values in Burmese Zawgyi encoding
Background: [Zawgyi](https://en.wikipedia.org/wiki/Zawgyi_font) is an
obsolete font encoding that is incompatible with proper Unicode.
Structurally, Zawgyi strings look like Unicode (they can be passed
around and stored as UTF-8, UTF-16, etc.), but when rendered, text
gets displayed as garbled characters unless the user happens to have a
non-standard font installed. With this non-standard font, the system
is able to render Zawgyi, but properly encoded Unicode strings look
broken. Also, because the Zawgyi encoding abuses codepoints intended
for Myanmar’s minority languages, installing a Zawgyi font breaks the
display of text in those minority languages.  The situation is a bit like with
ISO 8859 in the 1980s, but worse because Zawgyi text and fonts pretend
to be Unicode. Because of the structural similarity to Unicode,
detecting Zawgyi is non-trivial and can only be done probabilistically.
As of early 2025, Zawgyi is on the decline on the general Internet.
However, OpenStreetMap still contains thousands of objects with tag
values that are encoded in Burmese Zawgyi instead of proper Unicode.

This Osmose plugin makes use of Google’s open-source Zawgyi detector,
which uses Markov chains to estimate the likelihood of a Burmese
string being Zawgyi-encoded versus proper Unicode. The Osmose plugin
suggests fixes by converting mis-encoded strings from Zawgyi to
Unicode using the Unicode ICU library, which comes with a built-in
converter for this purpose.

Fixes https://github.com/osm-fr/osmose-backend/issues/2442.